pyproject.toml
[project]
name = "satelite"
version = "0.1.0"
description = "Distressed property lead finder for real estate wholesaling"
requires-python = ">=3.12"
dependencies = [
    "click>=8.0",
    "requests>=2.28",
    "aiohttp>=3.8",
    "aiofiles>=23.0",
    "Pillow>=10.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "pandas>=2.0",
    "rich>=13.0",
    "anthropic>=0.40",
    "streetlevel>=0.9",
    "pyyaml>=6.0",
]

[project.scripts]
satelite = "satelite.cli:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/satelite"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[project.optional-dependencies]
dev = ["pytest>=8.0", "pytest-asyncio>=0.23", "ruff>=0.4"]
